API and in-game business concept redesign beginning.
The idea is to replace standard models with rich models and move the main logic into them. This provides more flexibility than the logic in the controllers. The API in this commit won't work because of deleted controllers, but it will in the next commits.

The in-game business concept is being redesigned as well.